// src/app/protected/stocks/warehouses.interface.ts
export interface Warehouse {
    // moment: string;
    index: number;
    warehouseId: number;
    warehouseName: string;
    warehouseTel: string;
    warehouseAddress: string;
    warehouseComment: string;
    warehouseDefualt: boolean;
    warehouseStatus: boolean;
    warehouseLastupdate: any;
    warehouseCreated: any;

}